--- src/components/Post.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-import { useParams } from "react-router-dom";
-
-export const Post = () => {
-  const { id } = useParams();
-  return <p>Post: {id}</p>;
-};
